source/requirements.txt
aiobotocore==2.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
aiostalk==1.3.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
botocore==1.39.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
elasticsearch==7.15.0
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
fsspec==2025.2.0
google-auth==2.40.3
google-genai==1.30.0
greenlet==3.2.4
greenstalk==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.8
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.3.45
loguru==0.7.3
multidict==6.6.4
openai==1.99.9
orjson==3.11.2
ormsgpack==1.10.0
packaging==24.2
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
s3fs==2025.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
srsly==2.5.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.35.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
json_repair==0.49.0
python-multipart==0.0.20